scripts/requirements.txt
# Logo Processing Dependencies
rembg>=2.0.50
pillow>=10.0.0
numpy
